+++ projects/components/src/lib/checkbox/au-checkbox.directive.ts
@@ -52,6 +52,10 @@ export class AuCheckbox {
   readonly errors = input<readonly ValidationError.WithOptionalFieldTree[]>([]);
   readonly invalid = input(false);
 
+  readonly touched = input(false);
+  readonly dirty = input(false);
+  readonly submitting = input(false);
+
   readonly disabled = input(false);
   readonly required = input(false);
   readonly showRequired = input(true);
@@ -87,6 +91,8 @@ export class AuCheckbox {
   readonly effectiveInvalid = effectiveInvalidWithField(this.formField, {
     invalid: () => this.invalid(),
     isInvalid: () => this.isInvalid(),
+    touched: () => this.touched(),
+    dirty: () => this.dirty(),
   });
   readonly isChecked = computed(() => this.effectiveChecked());